requirements.txt
certifi==2018.4.16
Django==2.0.5
djangorestframework==3.8.2
mock==2.0.0
pbr==4.1.0
pytz==2018.5
six==1.11.0
